fix(code): retours revue Copilot sur DocBreadcrumbs
- Extrait Category + categoryLabels dans un module léger
  site/src/data/categories.ts pour ne pas embarquer le tableau
  resources (~4500 lignes) dans le bundle des pages docs où
  DocBreadcrumbs est rendu. resources.ts réexporte les deux pour
  préserver la rétro-compatibilité des autres imports.
- Le fallback href ne se déclenche plus si item.linkUnlisted est
  true : Docusaurus a volontairement retiré le lien dans ce cas
  (catégorie marquée unlisted), on respecte cette intention plutôt
  que de réintroduire un lien.

Refs #173
